P2.5 S8-раунд 2: нити F61 закрыты; задача → 👀
Прекомпьют scanPoints вне замеряемого цикла (скан 14,5→13,4 мкс,
TestScanClasses на общей функции); конверсия int64 до вычитания в
polyCollinear; устаревший комментарий и подпись кейса. Реестр: S8
закрыт — 0 блокеров/мажоров (бюджеты: план 1/2, рестарты 0/2, ревью
реализации 2/3; F1–F61, ложных 0). КТ-3 зафиксирован в журнале;
статус P2.5 → 👀, ждёт ✅ владельца

+++ internal/data/zone.go
@@ -631,8 +631,8 @@ func polyBounds(nodes [][2]int32) (minX, maxX, minY, maxY int32) {
 func polyCollinear(nodes [][2]int32) bool {
 	for i := range nodes {
 		p1, p2, p3 := nodes[i], nodes[(i+1)%len(nodes)], nodes[(i+2)%len(nodes)]
-		cross := int64(p2[0]-p1[0])*int64(p3[1]-p1[1]) -
-			int64(p2[1]-p1[1])*int64(p3[0]-p1[0])
+		cross := (int64(p2[0])-int64(p1[0]))*(int64(p3[1])-int64(p1[1])) -
+			(int64(p2[1])-int64(p1[1]))*(int64(p3[0])-int64(p1[0]))
 		if cross != 0 {
 			return false
 		}
